Create write a program to print the number of times the first string S1 appears in the second string S2 as a substring.
Given two strings S1 and S2, write a program to print the number of times the first string S1 appears in the second string S2 as a substring. 
For example if S1 is “atg” and S2 is “batgccatgatatga”, then the program should print 3. Program should work correctly
for all possible inputs. DO NOT use any of the library functions defined under strings described above.