Create Write a function which sorts an array x[ ] of n integers in ascending order of their absolute distances with a given number z
The absolute distance between two integers x1 and x2 is given by | x2 – x1 |. Write a function which sorts an array x[ ] of n integers in ascending order of their absolute distances with a given number z. For example, given x[ ] = {9, 1, 12, 4, 2} and z = 6, the sorted array will be x[ ] = {4, 9, 2, 1, 12}. Note that 4 is closest to 6, and 12 is farthest from 6, in terms of absolute distances. The function will
have the following prototype:
void dist_sort( int x[ ], int n, int z ) ;
Specify the time and space complexity in the report